log_reader: minor refactor of log file reading
This patch removes one goroutine that was in charge of reading a file
line by line in favor of pushing this functionality into the consumer. I
thought that between this goroutine and the consumer goroutine that
consumes lines and produces log entries, there were too many goroutines.
Also, starting off by splitting the input file into lines is not handy
for parsing CockroachDB log files (see next commit).

This patch also ensures that the input file is closed. Before, the file
was technically leaking - nobody was closing it. There is a finalizer
that does close the file at some random GC point, but "you're not
supposed to rely on that".