components: don't bubble up series and interactions
This patch makes it such that the components only look for data series
and interactions in their direct children in the template, and not
recursively.

Consider the following template
```
  <weighted-tree>

    <data-series>
      <query>
        <string>stacks.tree</string>
      </query>

      <interactions>
        <reaction target="data-series" type="fetch">
        ...
        </reaction>
      </interactions>
    </data-series>

  </weighted-tree>
```

Before this patch, the `interactions` specified under `<data-series>`
would bubble up to the `<weighted-tree>` component, causing it to
complain that it doesn't support the "data-series" target.
The error would go away if you add
```
<interactions></interactions>
```
directly under `<weighted-tree>`.