*: use a node's path to maintain object constancy in WeightedTree
Before this patch, the weighted tree component would map
RenderedTreeNodes to DOM elements essentially randomly every time a new
tree was received, when performing D3 data binding. Basically, a DOM
element previously corresponding to a node n1 would be mapped to some
other node n2 whenever the tree changed at all.

This patch makes it so that, if the backend cooperates, the component
will bind a tree node to the same DOM element as it was previously bound
to - assuming that the node was already part of the loaded tree. This
works by using the node's path as a node's identity and mapping nodes to
DOM elements according to this identity. So, if we just got a node nNew
with path p1, and a node nOld previously existed with the same path
(i.e. morally, our node was part of the tree previously loaded),
then nNew will be mapped to the DOM element that nOld had been mapped
to.

I believe this is generally what D3 wants one to do; without this logic
it's quite confusing to think about how D3 selections work. It's
supposed to result in:
a) fewer DOM changes (although in our case I'm not sure if that
currently applies to us given how we manipulate the DOM)
b) nice-looking visual transitions from the old tree to the new tree.
We don't quite seem to get nice transitions currently even with this
patch yet; but I think that should be doable.