logviz/client: dedup the logviz-client-core dep
This patch makes it so that, when building the logviz app, a single copy
of the traceviz-client-core package/library is used in case the app
imports logviz-client-core.

This is a bit of a theoretical concern given the current state of the
code; at the moment, logviz does not directly import
traceviz-client-core. But surely the app will import it in the future -
e.g. as soon as it wants to implement a new component that uses
interactions or such. More importantly, people copying the logviz code
to serve as a skeleton for their own apps will surely import
traceviz-client-core, and this patch will save them a lot of headache.

Without this patch, if logviz were to import t-c-c, the app would end up
with two copies it (one copy from the direct import, and one copy from
traceviz importing t-c-c). Moreover, classes from t-c-c are part of
traceviz's public interface (e.g. implementations of Value). Herein lies
the tragedy: if logviz were to create a StringValue using its copy of
t-c-c, and then pass it to traceviz which passes it to its own copy of
t-c-c, things break in extremely-hard-to-debug ways. JS is a dynamic,
duck-typed language, so not *everything* breaks, but things like
`instanceof` do.

The official solution for these duplicate dep issues is to make t-c-c a
"peer dependency" of traceviz, instead of a regular dependency. However,
that fails to solve the problem if the reference between the libraries
is a local reference created by `npm link` - like it currently is.
Instead, this patch uses webpack to make sure that, in the bundle
ultimately produced, everybody uses a single copy of the library.

[1] discussed the issue in more detail.

[1] https://penx.medium.com/managing-dependencies-in-a-node-package-so-that-they-are-compatible-with-npm-link-61befa5aaca7