data-table: allow user to inject styling
This is a bit out of left field; I want to see if it'll fly.
This patch allows users to inject CSS styling into the data table. I've
immediately needed this in order to override the `width: 100%` that the
data-table's styling specifies. But for a visual component like the
table, I think we generally want users to be able to specify any set of
styles they want; it feels like a component whose styling should be
highly configurable.

Surprisingly, figuring out exactly how to allow this kind of
customization on the data-table wasn't trivial at all. The official docs
have some very timid guidance ([1]). From what I can tell, we're in a
bit of a "there's a deprecated way, a not-ready-yet way"; this patch
implements the not-ready-yet way. The deprecated way involves using the
::ng-deep CSS selector to allow a parent component to pierce through the
"view encapsulation" that Angular does in order to keep a component's
style local to that component. Browsers are retiring support for this,
though.
Another way is to use the `part="<identifier>"` HTML attribute inside
the data-table, in order to provide a supported way for other components
to inject styling into this component's shadow dom. But that only works
if the component uses `encapsulation: ViewEncapsulation.ShadowDom`,
which is not the default because it's a relatively new browser feature.
I've switched to using it, though; all browsers do support it these
days. This post ([2]) helped with figuring out how to put it together.

With this patch, a parent component can specify something like this:
  styles: [`
    data-table::part(table) {background: #f4f4f4; border: 1px solid #dcdcdc; width: unset; min-width: 350px;}
    `]

[1] https://angular.io/guide/component-styles#authoring-a-component-to-support-customization
[2] https://kevinkreuzer.medium.com/the-ultimate-guide-on-how-to-create-customizable-angular-components-3eb9794bf86f