components: export the weighted tree targets and actions
Export the action targets and types by making them public static fields
of the WeightedTreeComponent class (thus also scoping them to the tree).

This is the minimum that I need for allowing client code to
programatically add event handlers to the WeightedTree. It looks like
this:

```
class MyComponent {
  @ViewChild(WeightedTreeComponent) weightedTree:  WeightedTreeComponent | undefined;

  ngAfterViewInit() {
    this.weightedTree!.interactionsDir?.get().withAction(
      new Action(WeightedTreeComponent.NODE, WeightedTreeComponent.CTRL_CLICK, new Call(this.onNodeCtrlClick)));
  }

  onNodeCtrlClick(localState: ValueMap): void {...}
}

class Call extends Update {
  private handler: (localState: ValueMap) => void;

  constructor(handler:((localState: ValueMap) => void)) {
    super();
    this.handler = handler
  }

  override update(localState?: ValueMap | undefined) {
    this.handler(localState!)
  }

  get autoDocument(): string {...}
}
```